# OPENAPI_SWAGGER_SPEC.yaml
openapi: 3.0.0
info:
  title: Asset Management Enterprise SaaS API
  description: Complete API documentation for Asset Management system with enterprise features
  version: 1.0.0
  contact:
    name: API Support
    url: https://example.com/support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8085/api/v1
    description: Local Development
  - url: https://api.example.com/api/v1
    description: Production
paths:
...
  /compliance/bog/controls:
    post:
      summary: Create or update (upsert) a BOG control by directiveRef
      operationId: upsertBogControl
      tags:
        - BOG Compliance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BogControlRequest'
      responses:
        '200':
          description: Control created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BogControlResponse'
        '400':
          description: Missing directiveRef or invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient authority
  /compliance/bog/controls/{id}/status:
    patch:
      summary: Update the status of a BOG control
      operationId: updateBogControlStatus
      tags:
        - BOG Compliance
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Control ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [NOT_IMPLEMENTED, PARTIAL, IMPLEMENTED, NOT_APPLICABLE]
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BogControlResponse'
        '400':
          description: Missing or invalid status
        '401':
          description: Not authenticated
        '403':
          description: Insufficient authority
        '404':
          description: Control not found
...
  /compliance/controls:
    get:
      summary: List Controls
      operationId: listCompliance_controls
      tags:
          - Compliance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of Controls
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
        '401':
          description: Unauthorized
...
  /compliance/controls/{id}:
    delete:
      summary: Delete Control
      operationId: deleteCompliance_controls
      tags:
          - Compliance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Control UUID
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BogControlRequest:
      type: object
      required:
        - directiveRef
      properties:
        directiveRef:
          type: string
          example: ICT.4.2
          description: BOG directive reference used as upsert key
        requirement:
          type: string
          example: Implement endpoint detection and response on all devices
        status:
          type: string
          enum: [NOT_IMPLEMENTED, PARTIAL, IMPLEMENTED, NOT_APPLICABLE]
          example: PARTIAL
        gapDescription:
          type: string
          example: EDR deployed on 60% of endpoints
        remediationPlan:
          type: string
          example: Complete rollout by Q3 2025
        targetDate:
          type: string
          format: date-time
          example: "2025-09-30T00:00:00Z"
        evidenceUrl:
          type: string
          example: https://docs.internal/edr-rollout
    BogControlResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        directiveRef:
          type: string
          example: ICT.4.2
        status:
          type: string
          enum: [NOT_IMPLEMENTED, PARTIAL, IMPLEMENTED, NOT_APPLICABLE]
...
    Error:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...